pyproject.toml
[project]
name = "pg-health"
version = "0.1.0"
description = "PostgreSQL health check and optimization tool"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.29",
    "typer>=0.9",
    "rich>=13",
    "pydantic>=2",
    "python-dotenv>=1",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "jinja2>=3",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
pg-health = "pg_health.cli:app"

[tool.ruff]
line-length = 100
